requirements.txt
# ===== Core Frameworks =====
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.2.0
google-adk>=0.3.0

# ===== LLM Providers =====
groq>=0.9.0
openai>=1.50.0
google-generativeai>=0.8.0
anthropic>=0.35.0

# ===== Observability =====
arize-phoenix>=5.0.0
openinference-instrumentation-langchain>=0.1.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0

# ===== Data Validation & Schemas =====
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ===== RAG & Vector Stores (Week 5) =====
faiss-cpu>=1.8.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ===== Context Graphs (Week 5) =====
neo4j>=5.25.0
graphiti-core>=0.3.0

# ===== API & HTTP =====
requests>=2.32.0
httpx>=0.27.0
aiohttp>=3.10.0

# ===== Error Handling & Resilience =====
tenacity>=9.0.0

# ===== Cost Tracking =====
tiktoken>=0.7.0

# ===== MCP (Week 7) =====
mcp>=1.0.0

# ===== Testing (Week 9) =====
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0

# ===== Deployment (Week 9) =====
fastapi>=0.115.0
uvicorn>=0.31.0

# ===== Utilities =====
python-dotenv>=1.0.0
rich>=13.9.0
tqdm>=4.66.0
pytz>=2024.1
